refactor: 기존 상속 방식 대신에 애노테이션 방식으로 전면 변경
