+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -13,12 +13,22 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Controller
-public class UserController {
+public class LoginController {
 
-    private static final Logger log = LoggerFactory.getLogger(UserController.class);
+    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
+
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView get(final HttpServletRequest request, final HttpServletResponse response) {
+        return UserSession.getUserFrom(request.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return new ModelAndView(new JspView("redirect:/index.jsp"));
+                })
+                .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
 
     @RequestMapping(value = "/login", method = RequestMethod.POST)
-    public ModelAndView login(final HttpServletRequest request, final HttpServletResponse response) {
+    public ModelAndView post(final HttpServletRequest request, final HttpServletResponse response) {
         if (UserSession.isLoggedIn(request.getSession())) {
             return new ModelAndView(new JspView("redirect:/index.jsp"));
         }
@@ -38,22 +48,5 @@ public class UserController {
             return new ModelAndView(new JspView("redirect:/index.jsp"));
         }
         return new ModelAndView(new JspView("redirect:/401.jsp"));
-    }
-
-    @RequestMapping(value = "/logout", method = RequestMethod.POST)
-    public ModelAndView logout(final HttpServletRequest request, final HttpServletResponse response) {
-        final var session = request.getSession();
-        session.removeAttribute(UserSession.SESSION_KEY);
-        return new ModelAndView(new JspView("redirect:/index.jsp"));
-    }
-
-    @RequestMapping(value = "/register", method = RequestMethod.POST)
-    public ModelAndView register(final HttpServletRequest request, final HttpServletResponse response) {
-        final var user = new User(2,
-                request.getParameter("account"),
-                request.getParameter("password"),
-                request.getParameter("email"));
-        InMemoryUserRepository.save(user);
-        return new ModelAndView(new JspView("redirect:/index.jsp"));
     }
 }
