+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,0 +1,30 @@
+package com.techcourse.controller;
+
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.view.JspView;
+import com.techcourse.domain.User;
+import com.techcourse.repository.InMemoryUserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Controller
+public class RegisterController {
+
+    @RequestMapping(value = "/register", method = RequestMethod.GET)
+    public ModelAndView get(final HttpServletRequest request, final HttpServletResponse response) {
+        return new ModelAndView(new JspView("/register.jsp"));
+    }
+
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public ModelAndView post(final HttpServletRequest request, final HttpServletResponse response) {
+        final var user = new User(2,
+                request.getParameter("account"),
+                request.getParameter("password"),
+                request.getParameter("email"));
+        InMemoryUserRepository.save(user);
+        return new ModelAndView(new JspView("redirect:/index.jsp"));
+    }
+}
